requirements.txt
torch>=1.0
torchvision
lmdb
pretrainedmodels
bootstrap.pytorch
tqdm
werkzeug
